TimeSeriesBean upgraded and DataTable added to dtw_update.jsp
Now, any sort of data structures (note the plural form) can be
transferred from the back-end to the front-end directly, without
converting them to Strings. This was made possible by adding
List<Object> resultObjects to the TimeSeriesBean class. The results
string was still retained for legacy purposes. :-P

In other news, relevant data for the DTW algorithm is displayed in a
data table in the GUI. Enjoy. There are some bugs related to SAX, but
those will (hopefully) be fixed soon. :-P